+++ examples/node.js/src/bytecode_call.ts
@@ -3,8 +3,8 @@ import {
   decodeFunctionReturn,
   encodeDeployData,
   encodeFunctionData,
-} from "@delvtech/drift";
-import { testToken } from "@delvtech/drift/testing";
+} from "@gud/drift";
+import { testToken } from "@gud/drift/testing";
 
 const rpcUrl = process.env.RPC_URL;
 const drift = createDrift({ rpcUrl });