+++ packages/drift-web3/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @delvtech/drift-web3
+# @gud/drift-web3
+
+## 1.0.2
+
+### Patch Changes
+
+- 5bbd2ed: Renamed packages from `@delvtech` to `@gud` scope.
 
 ## 1.0.1