+++ packages/drift/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @delvtech/drift
+# @gud/drift
+
+## 1.0.2
+
+### Patch Changes
+
+- 5bbd2ed: Renamed packages from `@delvtech` to `@gud` scope.
 
 ## 1.0.1